Clean up headers in nodewarden
