+++ ydb/core/blobstorage/nodewarden/bind_queue.h
@@ -1,6 +1,12 @@
 #pragma once
 
-#include "defs.h"
+#include <util/generic/hash.h>
+#include <util/generic/hash_set.h>
+#include <util/random/random.h>
+#include <util/string/builder.h>
+#include <ydb/library/yverify_stream/yverify_stream.h>
+
+#include <library/cpp/time_provider/monotonic.h>
 
 namespace NKikimr::NStorage {
