--- ydb/core/blobstorage/nodewarden/defs.h
@@ -1,66 +1,0 @@
-#pragma once
-
-#include <ydb/core/blobstorage/defs.h>
-
-#include <ydb/library/actors/core/log.h>
-#include <ydb/library/actors/core/actor_bootstrapped.h>
-#include <ydb/library/actors/core/hfunc.h>
-#include <ydb/library/actors/core/mailbox_queue_revolving.h>
-#include <ydb/library/actors/core/invoke.h>
-#include <ydb/library/actors/core/io_dispatcher.h>
-
-#include <ydb/library/services/services.pb.h>
-#include <ydb/core/protos/config.pb.h>
-
-#include <ydb/core/blobstorage/nodewarden/group_stat_aggregator.h>
-#include <ydb/core/base/appdata_fwd.h>
-#include <ydb/core/base/tablet_pipe.h>
-#include <ydb/core/base/tablet_resolver.h>
-#include <ydb/core/base/blobstorage_common.h>
-#include <ydb/core/blob_depot/agent/agent.h>
-#include <ydb/core/node_whiteboard/node_whiteboard.h>
-#include <ydb/core/blobstorage/base/blobstorage_events.h>
-#include <ydb/core/blobstorage/vdisk/vdisk_actor.h>
-#include <ydb/core/blobstorage/lwtrace_probes/blobstorage_probes.h>
-#include <ydb/core/blobstorage/incrhuge/incrhuge_keeper.h>
-#include <ydb/core/blobstorage/pdisk/blobstorage_pdisk_tools.h>
-#include <ydb/core/blobstorage/vdisk/vdisk_services.h>
-#include <ydb/core/blobstorage/dsproxy/dsproxy.h>
-#include <ydb/core/blobstorage/dsproxy/dsproxy_nodemon.h>
-#include <ydb/core/blobstorage/dsproxy/dsproxy_nodemonactor.h>
-#include <ydb/core/blobstorage/dsproxy/mock/dsproxy_mock.h>
-#include <ydb/core/blobstorage/vdisk/common/vdisk_config.h>
-#include <ydb/core/blobstorage/pdisk/blobstorage_pdisk_drivemodel_db.h>
-#include <ydb/core/blobstorage/pdisk/blobstorage_pdisk_factory.h>
-#include <ydb/core/mind/bscontroller/group_mapper.h>
-#include <ydb/core/mind/bscontroller/group_geometry_info.h>
-#include <ydb/core/util/log_priority_mute_checker.h>
-#include <ydb/core/util/stlog.h>
-#include <ydb/library/pdisk_io/sector_map.h>
-
-#include <library/cpp/lwtrace/mon/mon_lwtrace.h>
-
-#include <google/protobuf/text_format.h>
-#include <google/protobuf/util/json_util.h>
-
-#include <library/cpp/digest/crc32c/crc32c.h>
-#include <ydb/library/actors/interconnect/interconnect.h>
-#include <library/cpp/openssl/crypto/sha.h>
-#include <library/cpp/json/json_value.h>
-
-#include <util/folder/dirut.h>
-#include <util/folder/tempdir.h>
-#include <util/generic/algorithm.h>
-#include <util/generic/queue.h>
-#include <util/generic/set.h>
-#include <util/random/entropy.h>
-#include <util/stream/file.h>
-#include <util/stream/input.h>
-#include <util/string/escape.h>
-#include <util/string/printf.h>
-#include <util/system/backtrace.h>
-#include <util/system/condvar.h>
-#include <util/system/filemap.h>
-#include <util/system/file.h>
-#include <util/system/mutex.h>
-#include <util/system/thread.h>
