+++ ydb/core/blobstorage/nodewarden/node_warden_events.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include "defs.h"
-
+#include <ydb/core/base/blobstorage.h>
+#include <ydb/core/blobstorage/pdisk/blobstorage_pdisk_defs.h>
+#include <ydb/core/protos/blobstorage_config.pb.h>
 #include <ydb/core/protos/blobstorage_distributed_config.pb.h>
 
 namespace NKikimr::NStorage {
