+++ ydb/core/blobstorage/nodewarden/node_warden_impl.cpp
@@ -1,16 +1,149 @@
+#include "node_warden.h"
+#include "node_warden_events.h"
 #include "node_warden_impl.h"
-#include "distconf.h"
 
 #include <ydb/core/blobstorage/crypto/secured_block.h>
+#include <ydb/core/blobstorage/dsproxy/dsproxy.h>
 #include <ydb/core/blobstorage/dsproxy/dsproxy_request_reporting.h>
+#include <ydb/core/blobstorage/dsproxy/dsproxy_nodemonactor.h>
 #include <ydb/core/blobstorage/pdisk/drivedata_serializer.h>
+#include <ydb/core/blobstorage/vdisk/repl/blobstorage_replbroker.h>
 #include <ydb/library/pdisk_io/file_params.h>
 #include <ydb/core/base/nameservice.h>
-
 #include <ydb/core/protos/key.pb.h>
+
+#include <library/cpp/lwtrace/mon/mon_lwtrace.h>
 
 using namespace NKikimr;
 using namespace NStorage;
+
+TNodeWarden::TNodeWarden(const TIntrusivePtr<TNodeWardenConfig> &cfg)
+    : Cfg(cfg)
+    , EnablePutBatching(Cfg->FeatureFlags.GetEnablePutBatchingForBlobStorage(), false, true)
+    , EnableVPatch(Cfg->FeatureFlags.GetEnableVPatch(), false, true)
+    , EnableLocalSyncLogDataCutting(0, 0, 1)
+    , EnableSyncLogChunkCompressionHDD(1, 0, 1)
+    , EnableSyncLogChunkCompressionSSD(0, 0, 1)
+    , MaxSyncLogChunksInFlightHDD(10, 1, 1024)
+    , MaxSyncLogChunksInFlightSSD(10, 1, 1024)
+    , DefaultHugeGarbagePerMille(300, 1, 1000)
+    , MaxCommonLogChunksHDD(200, 1, 1'000'000)
+    , MaxCommonLogChunksSSD(200, 1, 1'000'000)
+    , CostMetricsParametersByMedia({
+        TCostMetricsParameters{200},
+        TCostMetricsParameters{50},
+        TCostMetricsParameters{32},
+    })
+    , SlowDiskThreshold(2'000, 1, 1'000'000)
+    , PredictedDelayMultiplier(1'000, 1, 1000)
+    , LongRequestThresholdMs(50'000, 1, 1'000'000)
+    , LongRequestReportingDelayMs(60'000, 1, 1'000'000)
+    , MaxNumOfSlowDisks(2, 1, 2)
+{
+    Y_ABORT_UNLESS(Cfg->BlobStorageConfig.GetServiceSet().AvailabilityDomainsSize() <= 1);
+    AvailDomainId = 1;
+    for (const auto& domain : Cfg->BlobStorageConfig.GetServiceSet().GetAvailabilityDomains()) {
+        AvailDomainId = domain;
+    }
+    if (Cfg->DomainsConfig) {
+        for (const auto& ssconf : Cfg->DomainsConfig->GetStateStorage()) {
+            BuildStateStorageInfos(ssconf, StateStorageInfo, BoardInfo, SchemeBoardInfo);
+            StateStorageProxyConfigured = true;
+        }
+    }
+}
+
+STATEFN(TNodeWarden::StateOnline) {
+    switch (ev->GetTypeRewrite()) {
+        fFunc(TEvBlobStorage::TEvPut::EventType, HandleForwarded);
+        fFunc(TEvBlobStorage::TEvGet::EventType, HandleForwarded);
+        fFunc(TEvBlobStorage::TEvBlock::EventType, HandleForwarded);
+        fFunc(TEvBlobStorage::TEvPatch::EventType, HandleForwarded);
+        fFunc(TEvBlobStorage::TEvDiscover::EventType, HandleForwarded);
+        fFunc(TEvBlobStorage::TEvRange::EventType, HandleForwarded);
+        fFunc(TEvBlobStorage::TEvCollectGarbage::EventType, HandleForwarded);
+        fFunc(TEvBlobStorage::TEvStatus::EventType, HandleForwarded);
+        fFunc(TEvBlobStorage::TEvAssimilate::EventType, HandleForwarded);
+        fFunc(TEvBlobStorage::TEvBunchOfEvents::EventType, HandleForwarded);
+        fFunc(TEvRequestProxySessionsState::EventType, HandleForwarded);
+
+        cFunc(TEvPrivate::EvGroupPendingQueueTick, HandleGroupPendingQueueTick);
+
+        hFunc(NIncrHuge::TEvIncrHugeInit, HandleIncrHugeInit);
+
+        hFunc(TEvInterconnect::TEvNodeInfo, Handle);
+
+        hFunc(TEvTabletPipe::TEvClientConnected, Handle);
+        hFunc(TEvTabletPipe::TEvClientDestroyed, Handle);
+
+        hFunc(NPDisk::TEvSlayResult, Handle);
+
+        hFunc(TEvRegisterPDiskLoadActor, Handle);
+
+        hFunc(TEvStatusUpdate, Handle);
+        hFunc(TEvBlobStorage::TEvDropDonor, Handle);
+        hFunc(TEvBlobStorage::TEvAskRestartVDisk, Handle);
+        hFunc(TEvBlobStorage::TEvAskWardenRestartPDisk, Handle);
+        hFunc(TEvBlobStorage::TEvNotifyWardenPDiskRestarted, Handle);
+
+        hFunc(TEvGroupStatReport, Handle);
+
+        hFunc(TEvBlobStorage::TEvControllerNodeServiceSetUpdate, Handle);
+        hFunc(TEvBlobStorage::TEvUpdateGroupInfo, Handle);
+        hFunc(TEvBlobStorage::TEvControllerUpdateDiskStatus, Handle);
+        hFunc(TEvBlobStorage::TEvControllerGroupMetricsExchange, Handle);
+        hFunc(TEvPrivate::TEvSendDiskMetrics, Handle);
+        hFunc(TEvPrivate::TEvUpdateNodeDrives, Handle);
+        hFunc(NMon::TEvHttpInfo, Handle);
+        cFunc(NActors::TEvents::TSystem::Poison, PassAway);
+
+        hFunc(TEvBlobStorage::TEvControllerScrubQueryStartQuantum, Handle);
+        hFunc(TEvBlobStorage::TEvControllerScrubStartQuantum, Handle);
+        hFunc(TEvBlobStorage::TEvControllerScrubQuantumFinished, Handle);
+
+        hFunc(TEvents::TEvInvokeResult, Handle);
+
+        hFunc(TEvNodeWardenQueryGroupInfo, Handle);
+        hFunc(TEvNodeWardenQueryStorageConfig, Handle);
+        hFunc(TEvNodeWardenStorageConfig, Handle);
+        fFunc(TEvents::TSystem::Unsubscribe, HandleUnsubscribe);
+
+        // proxy requests for the NodeWhiteboard to prevent races
+        hFunc(NNodeWhiteboard::TEvWhiteboard::TEvBSGroupStateUpdate, Handle);
+
+        hFunc(TEvBlobStorage::TEvControllerConfigResponse, Handle);
+
+        cFunc(TEvPrivate::EvReadCache, HandleReadCache);
+        fFunc(TEvPrivate::EvGetGroup, HandleGetGroup);
+
+        fFunc(TEvBlobStorage::EvNodeConfigPush, ForwardToDistributedConfigKeeper);
+        fFunc(TEvBlobStorage::EvNodeConfigReversePush, ForwardToDistributedConfigKeeper);
+        fFunc(TEvBlobStorage::EvNodeConfigUnbind, ForwardToDistributedConfigKeeper);
+        fFunc(TEvBlobStorage::EvNodeConfigScatter, ForwardToDistributedConfigKeeper);
+        fFunc(TEvBlobStorage::EvNodeConfigGather, ForwardToDistributedConfigKeeper);
+        fFunc(TEvBlobStorage::EvNodeConfigInvokeOnRoot, ForwardToDistributedConfigKeeper);
+        fFunc(TEvBlobStorage::EvNodeWardenDynamicConfigSubscribe, ForwardToDistributedConfigKeeper);
+        fFunc(TEvBlobStorage::EvNodeWardenDynamicConfigPush, ForwardToDistributedConfigKeeper);
+
+        hFunc(TEvNodeWardenQueryBaseConfig, Handle);
+        hFunc(TEvNodeConfigInvokeOnRootResult, Handle);
+
+        fFunc(TEvents::TSystem::Gone, HandleGone);
+
+        hFunc(TEvNodeWardenReadMetadata, Handle);
+        hFunc(TEvNodeWardenWriteMetadata, Handle);
+        hFunc(TEvPrivate::TEvDereferencePDisk, Handle);
+
+        default:
+            EnqueuePendingMessage(ev);
+            break;
+    }
+
+    if (VDiskStatusChanged) {
+        SendDiskMetrics(false);
+        VDiskStatusChanged = false;
+    }
+}
 
 void TNodeWarden::RemoveDrivesWithBadSerialsAndReport(TVector<NPDisk::TDriveData>& drives, TStringStream& details) {
     // Serial number's size definitely won't exceed this number of bytes.
