+++ ydb/core/blobstorage/nodewarden/node_warden_vdisk.cpp
@@ -1,6 +1,8 @@
+#include "node_warden.h"
 #include "node_warden_impl.h"
 
 #include <ydb/core/blobstorage/crypto/default.h>
+#include <ydb/core/blobstorage/vdisk/vdisk_actor.h>
 
 #include <util/string/split.h>
